backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
dash==3.0.3
dataclasses-json==0.6.7
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx==0.2.4
easyocr==1.7.2
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fdb==2.0.2
filelock==3.16.1
firebird-base==1.8.0
firebird-driver==1.10.6
Flask==3.0.3
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
ghostscript==0.7
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.31
lazy_loader==0.4
llama_cpp_python==0.3.8
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.1.2
oauthlib==3.2.2
openai==1.74.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
propcache==0.3.1
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.9.0
PyMsgBox==1.0.9
PyMuPDF==1.25.1
PyMySQL==1.1.1
pyodbc==5.2.0
pypdf==5.4.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-pkcs12==1.25
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.25.0
sentence-transformers==4.1.0
setuptools==75.2.0
shapely==2.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlalchemy-access==2.0.3
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
virtualenv==20.27.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
wsproto==1.2.0
wxPython==4.2.2
yarl==1.19.0
zeep==4.3.1
zipp==3.21.0
zstandard==0.23.0
